finished loopcreator

+++ src/minimumsteinertree.cpp
@@ -72,7 +72,7 @@ void MinimumSteinerTree::run()
 {
     DM::Logger(DM::Standard) << "Setup Graph";
 
-    typedef adjacency_list < vecS, vecS, undirectedS, property<vertex_distance_t, int>, property < edge_weight_t, double > > Graph;
+    typedef adjacency_list < vecS, vecS, undirectedS, property<vertex_distance_t, double>, property < edge_weight_t, double > > Graph;
     typedef std::pair < int, int >E;
 
     this->sys = this->getData("Layout");
